Refactor code to replace instances of "container" with "section"
Fixed errors related to the use of the "container" element by replacing it with the more appropriate "section" element. The commit includes changes to three sections of the code (Berufliche, Berufsausbildung, and skils), where the "container" element was being used as a child of the "div" element. These changes were made to conform to HTML content model specifications, which state that "div" elements should only contain flow content, and clear some space lines